python/flask/pyproject.toml
[tool.poetry]
name = "spackle-example-flask"
version = "0.1.0"
description = ""
authors = ["Hunter Clarke <hunter@spackle.so>"]
license = "MIT"
readme = "README.md"
packages = [{include = "app"}]


[tool.poetry.dependencies]
python = "^3.9"
spackle-python = "^0.0.44b1"
Flask = "^2.2.3"
python-dotenv = "^1.0.0"
gunicorn = "^20.1.0"
setuptools = "^67.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
